Commit-9-ServiceClassAndTestsFullyImplemented

+++ src/main/java/Service/MessageService.java
@@ -19,7 +19,7 @@ public class MessageService {
 
     //User Story 3 Service Method 
     public Message createMessage(Message msg){
-        if(msg.getMessage_text().length()<=255 && !msg.getMessage_text().isEmpty() && acc_dao.checkAccountExistsByID(msg.getMessage_id())){
+        if(msg.getMessage_text().length()<=255 && !msg.getMessage_text().isEmpty() && acc_dao.checkAccountExistsByID(msg.getPosted_by())){
             Message msg_ret = msg_dao.createMessage(msg.getPosted_by(), msg.getMessage_text(), msg.getTime_posted_epoch());
             return msg_ret;
         }
@@ -38,12 +38,18 @@ public class MessageService {
 
     //User Story 6 Service Method  
     public Message deleteMessage(int id){
-        return msg_dao.deleteMessageByID(id);
+        if (getMessage(id)!=null){
+            return msg_dao.deleteMessageByID(id);
+        }
+        return null;
     }
 
     //User Story 7 Service Method 
     public Message updateMessage(int id, String msg_text){
-        return msg_dao.updateMessageByID(id, msg_text);
+        if (msg_text.length()<=255 && !msg_text.isEmpty() && getMessage(id)!=null){
+            return msg_dao.updateMessageByID(id, msg_text);
+        }
+        return null;
     }
 
     //User story 8 Service Method
